refactor: Initialize broker in daemon
The chain daemon -> dbusservice -> broker needed to pass configs through
multiple layers, which made the code hard to navigate and maintain.
