+++ cmd/authd-oidc/daemon/testdata/TestParseConfig/golden/do_not_fail_if_values_contain_a_single_template_delimiter
@@ -1,0 +1,4 @@
+DEFAULT: {}
+oidc:
+    client_id: <CLIENT_ID
+    issuer: https://ISSUER_URL>
