+++ cmd/authd-oidc/daemon/testdata/TestParseConfig/golden/successfully_parse_config_file
@@ -1,0 +1,4 @@
+DEFAULT: {}
+oidc:
+    client_id: client_id
+    issuer: https://issuer.url.com
