+++ cmd/authd-oidc/daemon/testdata/TestParseConfig/golden/successfully_parse_config_file_with_optional_values
@@ -1,0 +1,7 @@
+DEFAULT: {}
+oidc:
+    client_id: client_id
+    issuer: https://issuer.url.com
+users:
+    allowed_ssh_suffixes: '@issuer.url.com'
+    home_base_dir: /home
